Revert "RHEL: Fix virConnectGetMaxVcpus output"
This reverts commit fcec98bb80633bec6f4bc3de0ab75627c874d315.

Upstream KVM changed its soft limit for number of vCPUs to match the
number of physical CPUs on the host. This will make it into the kernel
in 9.1.0 and thus downstream qemu-kvm dropped their RHEL-only patch to
consider KVM soft limit as hard (otherwise no CPU overcommitment would
be possible). We need to do the same to avoid reporting wrong numbers.

https://bugzilla.redhat.com/show_bug.cgi?id=2095260

Signed-off-by: Jiri Denemark <jdenemar@redhat.com>